Rename bloom_lib.js to bloomSynphonyExtensions.js
