+++ src/BloomBrowserUI/bookEdit/toolbox/readers/libSynphony/bloomSynphonyExtensions.d.ts
