+++ src/BloomBrowserUI/bookEdit/toolbox/readers/readerSetup/ReaderSetup.jade
@@ -12,7 +12,7 @@ html
 		// script(src='/bloom/bookEdit/toolbox/decodableReader/libSynphony/underscore_min_152.js')
 		// script(src='/bloom/bookEdit/toolbox/decodableReader/libSynphony/jquery.text-markup.js')
 		// script(src='/bloom/bookEdit/toolbox/decodableReader/libSynphony/synphony_lib.js')
-		// script(src='/bloom/bookEdit/toolbox/decodableReader/libSynphony/bloom_lib.js')
+		// script(src='/bloom/bookEdit/toolbox/decodableReader/libSynphony/bloomSynphonyExtensions.js')
 		// script(src='/bloom/lib/localizationManager/localizationManager.js')
 		// script(src='/bloom/lib/jquery.i18n.custom.js')
 		// script(src='/bloom/lib/long-press/jquery.mousewheel.js')
